Conflicts:
	src/ptest/CycleTestBase.h
	src/ptest/CycleTestManger.cpp
	src/ptest/createminterkey.cpp
	src/ptest/createminterkey.h
	src/rpc/rpctx.cpp
	src/test/systestbase.h
	src/test/systestscript.cpp
	src/tx.cpp


Former-commit-id: a2e1a5c2f63e36f486972fd407266613d4958438

+++ src/noui.cpp
@@ -68,8 +68,8 @@ static void noui_BlockChanged(int64_t time,int64_t high,const uint256 &hash) {
 	if (CUIServer::HasConnection()) {
 		Object obj;
 		obj.push_back(Pair("type",     "blockchanged"));
-		obj.push_back(Pair("time",     time));
-		obj.push_back(Pair("high",     high));
+		obj.push_back(Pair("time",     (int)time));
+		obj.push_back(Pair("high",     (int)high));
 		obj.push_back(Pair("hash",     hash.ToString()));
 		CUIServer::Send(write_string(Value(std::move(obj)),true));
 	}
